requirements.txt
torch>=2.1.0
torchvision>=0.18.0
torchaudio>=2.1.0
numpy>=1.26.4
pandas>=2.2.2
matplotlib>=3.8.4
Pillow>=10.3.0
opencv-python>=4.9.0
transformers>=4.11.3
albumentations>=1.4.4
seaborn>=0.13.2
lightning>=2.2.4
wandb>=0.17.0
PyYAML>=6.0.1
monai>=1.3.1